requirements.txt
python-binance
discord.py
requests
aiohttp
pandas
numpy
pytrends
python-dotenv
nltk
vaderSentiment
mplfinance
matplotlib
fastapi
uvicorn[standard]

# Optional: try pandas-ta but may fail on Python 3.14
# pip install pandas-ta

# Optional: heavier, more accurate sentiment
# pip install transformers torch
